v1.105.6

+++ tests/features/steps/cli_testing/cli_projects.py
@@ -91,4 +91,4 @@ def step_impl(context, project_name):
 @behave.then(u'I wait "{seconds}"')
 @behave.when(u'I wait "{seconds}"')
 def step_impl(_, seconds):
-    time.sleep(int(seconds))
+    time.sleep(float(seconds))
